project/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "maildrop"
version = "0.1.0"
description = "CLI service for sorting IMAP email attachments into a local Nextcloud folder"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.3",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.5.0",
]

[project.scripts]
maildrop = "maildrop.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]